put enemy classes and MarkovAI class in same file, depricate constants file
