+++ universe.py
@@ -12,10 +12,11 @@ import pygame
 from pygame import Color
 from pygame.math import Vector2 as Vec2
 
+from enemy_ai import BulletEnemy, EnemyConfig, MarkovEnemy, MissileEnemy, RocketEnemy
 from physics import Disk, Particle, Pos, PosVel
 from profiler import global_profiler
 from projectiles import Missile
-from ship import BulletEnemy, EnemyConfig, MarkovEnemy, MissileEnemy, PlayerConfig, PlayerShip, RocketEnemy, ShipInput
+from ship import PlayerConfig, PlayerShip, ShipInput
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterator, Sequence
@@ -24,7 +25,7 @@ if TYPE_CHECKING:
     from projectiles import Bullet
     from ship import Ship
 
-from constants import GRAVITATIONAL_CONSTANT
+from physics import GRAVITATIONAL_CONSTANT
 
 PLANET_SIZE_PARAMETER = 5.9
 SIGMA_PLANET_RADIUS = 0.3
